% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallet.R
\name{mallet.topic.load}
\alias{mallet.topic.load}
\title{Load a topic from a file}
\usage{
mallet.topic.load(filename)
}
\arguments{
\item{filename}{The mallet topic file}
}
\description{
This function returns the topic model loaded from a file.
}
