% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallet.R
\name{mallet.topic.model.read}
\alias{mallet.topic.model.read}
\alias{mallet.topic.model.load}
\alias{mallet.topic.model.write}
\alias{mallet.topic.model.save}
\title{Load (read) and save (write) a topic from a file}
\usage{
mallet.topic.model.read(filename)

mallet.topic.model.load(filename)

mallet.topic.model.write(topic.model, filename)

mallet.topic.model.save(topic.model, filename)
}
\arguments{
\item{filename}{The mallet topic model file}

\item{topic.model}{A \code{cc.mallet.topics.RTopicModel} object created by \code{\link{MalletLDA}}.}
}
\description{
This function returns the topic model loaded from a file or stores a topic model to file.
}
