% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallet_jar.R
\name{mallet_jar}
\alias{mallet_jar}
\alias{mallet.jar}
\title{Return the mallet jar filename(s)}
\usage{
mallet_jar(full.names = FALSE)

mallet.jar(full.names = FALSE)
}
\arguments{
\item{full.names}{a logical value.
If TRUE, the directory path is prepended to the file names to give
a relative file path. If FALSE, the file name(s) (rather than paths)
are returned.}
}
\description{
Return the mallet jar filename(s)
}
\details{
Mallet is implemented as a jar-file in the mallet R package.
This function returns the file name and file path for that file(s)
}
