% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallet_stopwords.R
\name{mallet_stoplist_file_path}
\alias{mallet_stoplist_file_path}
\alias{mallet.stoplist.file.path}
\title{Return the file path to the mallet stoplists}
\usage{
mallet_stoplist_file_path(language = "en")

mallet.stoplist.file.path(language = "en")
}
\arguments{
\item{language}{language to return stoplist for. Defaults to engligs ([en]).}
}
\description{
Return the file path to the mallet stoplists
}
\details{
Returns the path to the mallet stop word list.
See [mallet_supported_stoplists()] for which stoplists that are included.
}
