% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallet_stopwords.R
\name{mallet_supported_stoplists}
\alias{mallet_supported_stoplists}
\alias{mallet.supported.stoplists}
\title{Mallet supported stoplists}
\usage{
mallet_supported_stoplists()

mallet.supported.stoplists()
}
\description{
Mallet supported stoplists
}
\details{
return vector with included stoplists
}
