% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallet.R
\name{save.mallet.instances}
\alias{save.mallet.instances}
\alias{load.mallet.instances}
\title{Load and save mallet instances from/to file}
\usage{
save.mallet.instances(instances, filename)

load.mallet.instances(filename)
}
\arguments{
\item{instances}{An instances object to save/write to file}

\item{filename}{The mallet topic model file}
}
\description{
This function returns the topic model loaded from a file.
}
