% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallet_state_io.R
\name{save.mallet.state}
\alias{save.mallet.state}
\title{Save a Mallet state to file}
\usage{
save.mallet.state(topic.model, state.file)
}
\arguments{
\item{topic.model}{a mallet \code{topic.model} object}

\item{state.file}{File path (.gz format) to store the mallet state file to.}
}
\description{
This function writes a current sampling state of mallet to file. 
The state contain hyperparameters \eqn{\alpha} and \eqn{\beta} together with topic indicators.

The state file can be read into R using the function
}
