% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallet.R
\docType{data}
\name{sotu}
\alias{sotu}
\title{State of the Union Adresses.}
\format{
A \code{\link[tibble]{tibble}} \code{data.frame} with 7625 rows and 3 variables:
\describe{
  \item{year}{Year of the adress.}
  \item{paragraph}{The paragraph of the address.}
  \item{text}{The address content.}
}
}
\source{
\url{https://en.wikipedia.org/wiki/State_of_the_Union}
}
\usage{
sotu
}
\description{
A dataset containing State of the Union Adresses by paragraph from 1945 to 2009.
}
\keyword{datasets}
