% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mallet.R
\docType{package}
\name{mallet-package}
\alias{mallet-package}
\title{An R Wrapper for the Mallet Topic Modeling Package}
\description{
This package provides an interface to the Java implementation of latent
Dirichlet allocation in the Mallet machine learning package. Mallet has many
functions, this wrapper focuses on the topic modeling sub-package written by
David Mimno. The package uses the \code{rJava} package to connect to a JVM.
}
\references{
The model, Latent Dirichlet allocation (LDA):
\cite{David M Blei, Andrew Ng, Michael Jordan. Latent Dirichlet Allocation. J. of Machine Learning Research, 2003.}

The Java toolkit:
\cite{Andrew Kachites McCallum. The Mallet Toolkit. 2002.}

Details of the fast sparse Gibbs sampling algorithm:
\cite{Limin Yao, David Mimno, Andrew McCallum. Streaming Inference for Latent Dirichlet Allocation. KDD, 2009.}

Hyperparameter optimization:
\cite{Hanna Wallach, David Mimno, Andrew McCallum. Rethinking LDA: Why Priors Matter. NIPS, 2010.}
}
